deprecate PolicySet
