+++ Sources/X509/Verifier/VerifierPolicy.swift
@@ -57,206 +57,3 @@ public enum PolicyEvaluationResult: Sendable {
     case meetsPolicy
     case failsToMeetPolicy(reason: String)
 }
-
-// TODO: Several enhancements.
-//
-// This type should properly be a variadic generic over `VerifierPolicy` to allow proper composition.
-// Additionally, we should add conditional Sendable, Equatable, and Hashable conformances as needed.
-// This will also allow equivalent conditional conformances on `Verifier`.
-public struct PolicySet: VerifierPolicy {
-    public let verifyingCriticalExtensions: [ASN1ObjectIdentifier]
-
-    @usableFromInline var policies: [any VerifierPolicy]
-
-    @inlinable
-    public init(policies: [any VerifierPolicy]) {
-        self.policies = policies
-
-        var extensions: [ASN1ObjectIdentifier] = []
-        extensions.reserveCapacity(policies.reduce(into: 0, { $0 += $1.verifyingCriticalExtensions.count }))
-
-        for policy in policies {
-            extensions.append(contentsOf: policy.verifyingCriticalExtensions)
-        }
-
-        self.verifyingCriticalExtensions = extensions
-    }
-
-    public mutating func chainMeetsPolicyRequirements(chain: UnverifiedCertificateChain) async -> PolicyEvaluationResult {
-        var policyIndex = self.policies.startIndex
-
-        while policyIndex < self.policies.endIndex {
-            switch await self.policies[policyIndex].chainMeetsPolicyRequirements(chain: chain) {
-            case .meetsPolicy:
-                ()
-            case .failsToMeetPolicy(reason: let reason):
-                return .failsToMeetPolicy(reason: reason)
-            }
-
-            self.policies.formIndex(after: &policyIndex)
-        }
-
-        return .meetsPolicy
-    }
-}
-
-public func makePolicy<Policy: VerifierPolicy>(
-    @PolicyBuilder _ body: () throws -> Policy
-) rethrows -> Policy {
-    try body()
-}
-
-extension Optional: VerifierPolicy where Wrapped: VerifierPolicy {
-    @inlinable
-    public var verifyingCriticalExtensions: [SwiftASN1.ASN1ObjectIdentifier] {
-        self?.verifyingCriticalExtensions ?? []
-    }
-    
-    @inlinable
-    public mutating func chainMeetsPolicyRequirements(chain: UnverifiedCertificateChain) async -> PolicyEvaluationResult {
-        await self?.chainMeetsPolicyRequirements(chain: chain) ?? .meetsPolicy
-    }
-}
-
-public struct AnyPolicy: VerifierPolicy {
-    @usableFromInline
-    var policy: any VerifierPolicy
-    
-    @inlinable
-    public init(_ policy: some VerifierPolicy) {
-        self.policy = policy
-    }
-    
-    @inlinable
-    public var verifyingCriticalExtensions: [SwiftASN1.ASN1ObjectIdentifier] {
-        policy.verifyingCriticalExtensions
-    }
-    
-    @inlinable
-    public mutating func chainMeetsPolicyRequirements(chain: UnverifiedCertificateChain) async -> PolicyEvaluationResult {
-        await policy.chainMeetsPolicyRequirements(chain: chain)
-    }
-}
-
-@resultBuilder
-public struct PolicyBuilder {
-    @usableFromInline
-    struct Empty: VerifierPolicy {
-        @inlinable
-        var verifyingCriticalExtensions: [SwiftASN1.ASN1ObjectIdentifier] { [] }
-        
-        @inlinable
-        init() {}
-        
-        @inlinable
-        mutating func chainMeetsPolicyRequirements(chain: UnverifiedCertificateChain) async -> PolicyEvaluationResult {
-            .meetsPolicy
-        }
-    }
-    
-    @inlinable
-    public static func buildBlock() -> some VerifierPolicy {
-        Empty()
-    }
-    
-    @usableFromInline
-    struct Tuple2<First: VerifierPolicy, Second: VerifierPolicy>: VerifierPolicy {
-        @usableFromInline
-        var first: First
-        
-        @usableFromInline
-        var second: Second
-        
-        @inlinable
-        init(first: First, second: Second) {
-            self.first = first
-            self.second = second
-        }
-        
-        @inlinable
-        var verifyingCriticalExtensions: [SwiftASN1.ASN1ObjectIdentifier] {
-            first.verifyingCriticalExtensions + second.verifyingCriticalExtensions
-        }
-        
-        @inlinable
-        mutating func chainMeetsPolicyRequirements(chain: UnverifiedCertificateChain) async -> PolicyEvaluationResult {
-            switch await first.chainMeetsPolicyRequirements(chain: chain) {
-            case .meetsPolicy:
-                break
-            case .failsToMeetPolicy(let reason):
-                return .failsToMeetPolicy(reason: reason)
-            }
-            
-            return await second.chainMeetsPolicyRequirements(chain: chain)
-        }
-    }
-    
-    @inlinable
-    public static func buildPartialBlock(first: some VerifierPolicy) -> some VerifierPolicy {
-        first
-    }
-    
-    @inlinable
-    public static func buildPartialBlock(accumulated: some VerifierPolicy, next: some VerifierPolicy) -> some VerifierPolicy {
-        Tuple2(first: accumulated, second: next)
-    }
-    
-    @inlinable
-    public static func buildOptional<Policy: VerifierPolicy>(_ component: Optional<Policy>) -> some VerifierPolicy {
-        component
-    }
-    
-    @inlinable
-    public static func buildBlock(_ components: some VerifierPolicy) -> some VerifierPolicy {
-        components
-    }
-    
-    @inlinable
-    public static func buildExpression(_ expression: some VerifierPolicy) -> some VerifierPolicy {
-        expression
-    }
-}
-
-
-
-
-
-
-// MARK: support dynamic policies
-extension PolicyBuilder {
-    @usableFromInline
-    internal enum Either<First: VerifierPolicy, Second: VerifierPolicy>: VerifierPolicy {
-        case first(First)
-        case second(Second)
-        
-        @inlinable
-        public var verifyingCriticalExtensions: [ASN1ObjectIdentifier] {
-            switch self {
-            case .first(let first): return first.verifyingCriticalExtensions
-            case .second(let second): return second.verifyingCriticalExtensions
-            }
-        }
-        
-        @inlinable
-        public mutating func chainMeetsPolicyRequirements(chain: UnverifiedCertificateChain) async -> PolicyEvaluationResult {
-            switch self {
-            case .first(var first):
-                defer { self = .first(first) }
-                return await first.chainMeetsPolicyRequirements(chain: chain)
-            case .second(var second):
-                defer { self = .second(second) }
-                return await second.chainMeetsPolicyRequirements(chain: chain)
-            }
-        }
-    }
-    
-    @inlinable
-    internal static func buildEither<First: VerifierPolicy, Second: VerifierPolicy>(first component: First) -> Either<First, Second> {
-        Either<First, Second>.first(component)
-    }
-    
-    @inlinable
-    internal static func buildEither<First: VerifierPolicy, Second: VerifierPolicy>(second component: Second) -> Either<First, Second> {
-        Either<First, Second>.second(component)
-    }
-}
